实体类数据类型更改

+++ src/main/java/com/lxw/videoworld/spider/MpdySourceDetailProcessor.java
@@ -79,7 +79,7 @@ public class MpdySourceDetailProcessor extends BasePhdyProcessor {
             sourceDetail.setTitle(title.trim());
         }
         if(!TextUtils.isEmpty(date) && date.length() == 8){
-            sourceDetail.setDate(Integer.valueOf(date));
+            sourceDetail.setDate(date);
         }
         if(imgUrl != null && imgUrl.size() > 0){
             sourceDetail.setImages(imgUrl.toString());
@@ -154,7 +154,7 @@ public class MpdySourceDetailProcessor extends BasePhdyProcessor {
                         if (!TextUtils.isEmpty(year) && year.length() > 4) {
                             year = year.substring(0, 4);
                         }
-                        sourceDetail.setYear(Integer.valueOf(year));
+                        sourceDetail.setYear(year);
                     } else {
                         String year1 = page.getHtml().css("div.bs-docs-section").regex("【年 &nbsp; &nbsp;　代】：(.*?)<").toString();
                         if (!TextUtils.isEmpty(year1)) {
@@ -163,7 +163,7 @@ public class MpdySourceDetailProcessor extends BasePhdyProcessor {
                             if (!TextUtils.isEmpty(year1) && year1.length() > 4) {
                                 year1 = year1.substring(0, 4);
                             }
-                            sourceDetail.setYear(Integer.valueOf(year1));
+                            sourceDetail.setYear(year1);
                         } else {
                             String year2 = page.getHtml().css("div.bs-docs-section").regex("年　　代(.*?)<").toString();
                             if (!TextUtils.isEmpty(year2)) {
@@ -172,7 +172,7 @@ public class MpdySourceDetailProcessor extends BasePhdyProcessor {
                                 if (!TextUtils.isEmpty(year2) && year2.length() > 4) {
                                     year2 = year2.substring(0, 4);
                                 }
-                                sourceDetail.setYear(Integer.valueOf(year2));
+                                sourceDetail.setYear(year2);
                             }
                         }
                     }
@@ -288,7 +288,7 @@ public class MpdySourceDetailProcessor extends BasePhdyProcessor {
                         imdb_score0 = imdb_score0.replaceAll("</font>", "");
                         String[] imdb_score = imdb_score0.trim().split("/");
                         if (imdb_score.length > 0 && !TextUtils.isEmpty(imdb_score[0])) {
-                            sourceDetail.setImdbScore(Float.valueOf(StringUtil.disposeField(imdb_score[0])));
+                            sourceDetail.setImdbScore(StringUtil.disposeField(imdb_score[0]));
                         }
                         if (imdb_score.length > 1 && !TextUtils.isEmpty(imdb_score[1])) {
                             sourceDetail.setImdbIntro(StringUtil.disposeField(imdb_score[1]));
@@ -299,7 +299,7 @@ public class MpdySourceDetailProcessor extends BasePhdyProcessor {
                             imdb_score1 = imdb_score1.replaceAll("</font>", "");
                             String[] imdb_score = imdb_score1.trim().split("/");
                             if (imdb_score.length > 0 && !TextUtils.isEmpty(imdb_score[0])) {
-                                sourceDetail.setImdbScore(Float.valueOf(StringUtil.disposeField(imdb_score[0])));
+                                sourceDetail.setImdbScore(StringUtil.disposeField(imdb_score[0]));
                             }
                             if (imdb_score.length > 1 && !TextUtils.isEmpty(imdb_score[1])) {
                                 sourceDetail.setImdbIntro(StringUtil.disposeField(imdb_score[1]));
@@ -315,7 +315,7 @@ public class MpdySourceDetailProcessor extends BasePhdyProcessor {
                         douban_score0 = douban_score0.replaceAll("</font>", "");
                         String[] douban_score = douban_score0.trim().split("/");
                         if (douban_score.length > 0 && !TextUtils.isEmpty(douban_score[0])) {
-                            sourceDetail.setDoubanScore(Float.valueOf(StringUtil.disposeField(douban_score[0])));
+                            sourceDetail.setDoubanScore(StringUtil.disposeField(douban_score[0]));
                         }
                         if (douban_score.length > 1 && !TextUtils.isEmpty(douban_score[1])) {
                             sourceDetail.setDoubanIntro(StringUtil.disposeField(douban_score[1]));
@@ -326,7 +326,7 @@ public class MpdySourceDetailProcessor extends BasePhdyProcessor {
                             douban_score1 = douban_score1.replaceAll("</font>", "");
                             String[] douban_score = douban_score1.trim().split("/");
                             if (douban_score.length > 0 && !TextUtils.isEmpty(douban_score[0])) {
-                                sourceDetail.setDoubanScore(Float.valueOf(StringUtil.disposeField(douban_score[0])));
+                                sourceDetail.setDoubanScore(StringUtil.disposeField(douban_score[0]));
                             }
                             if (douban_score.length > 1 && !TextUtils.isEmpty(douban_score[1])) {
                                 sourceDetail.setDoubanIntro(StringUtil.disposeField(douban_score[1]));
