+++ src/main/java/com/lxw/videoworld/spider/YgdySourceDetailProcessor.java
@@ -29,7 +29,7 @@ public class YgdySourceDetailProcessor extends BaseYgdyProcessor {
                 sourceDetail.setCategory(params[4]);
                 sourceDetail.setType(params[5]);
                 if (!TextUtils.isEmpty(params[6]) && params[6].length() == 8) {
-                    sourceDetail.setDate(Integer.valueOf(params[6]));
+                    sourceDetail.setDate(params[6]);
                 }
             } else {
 
@@ -188,7 +188,7 @@ public class YgdySourceDetailProcessor extends BaseYgdyProcessor {
                         if (!TextUtils.isEmpty(year) && year.length() > 4) {
                             year = year.substring(0, 4);
                         }
-                        sourceDetail.setYear(Integer.valueOf(year));
+                        sourceDetail.setYear(year);
                     } else {
                         String year1 = page.getHtml().css("div#Zoom").regex("【年 &nbsp; &nbsp;　代】：(.*?)<").toString();
                         if (!TextUtils.isEmpty(year1)) {
@@ -197,7 +197,7 @@ public class YgdySourceDetailProcessor extends BaseYgdyProcessor {
                             if (!TextUtils.isEmpty(year1) && year1.length() > 4) {
                                 year1 = year1.substring(0, 4);
                             }
-                            sourceDetail.setYear(Integer.valueOf(year1));
+                            sourceDetail.setYear(year1);
                         } else {
                             String year2 = page.getHtml().css("div#Zoom").regex("年　　代(.*?)<").toString();
                             if (!TextUtils.isEmpty(year2)) {
@@ -206,7 +206,7 @@ public class YgdySourceDetailProcessor extends BaseYgdyProcessor {
                                 if (!TextUtils.isEmpty(year2) && year2.length() > 4) {
                                     year2 = year2.substring(0, 4);
                                 }
-                                sourceDetail.setYear(Integer.valueOf(year2));
+                                sourceDetail.setYear(year2);
                             }
                         }
                     }
@@ -322,7 +322,7 @@ public class YgdySourceDetailProcessor extends BaseYgdyProcessor {
                         imdb_score0 = imdb_score0.replaceAll("</font>", "");
                         String[] imdb_score = imdb_score0.trim().split("/");
                         if (imdb_score.length > 0 && !TextUtils.isEmpty(imdb_score[0])) {
-                            sourceDetail.setImdbScore(Float.valueOf(StringUtil.disposeField(imdb_score[0])));
+                            sourceDetail.setImdbScore(StringUtil.disposeField(imdb_score[0]));
                         }
                         if (imdb_score.length > 1 && !TextUtils.isEmpty(imdb_score[1])) {
                             sourceDetail.setImdbIntro(StringUtil.disposeField(imdb_score[1]));
@@ -333,7 +333,7 @@ public class YgdySourceDetailProcessor extends BaseYgdyProcessor {
                             imdb_score1 = imdb_score1.replaceAll("</font>", "");
                             String[] imdb_score = imdb_score1.trim().split("/");
                             if (imdb_score.length > 0 && !TextUtils.isEmpty(imdb_score[0])) {
-                                sourceDetail.setImdbScore(Float.valueOf(StringUtil.disposeField(imdb_score[0])));
+                                sourceDetail.setImdbScore(StringUtil.disposeField(imdb_score[0]));
                             }
                             if (imdb_score.length > 1 && !TextUtils.isEmpty(imdb_score[1])) {
                                 sourceDetail.setImdbIntro(StringUtil.disposeField(imdb_score[1]));
@@ -349,7 +349,7 @@ public class YgdySourceDetailProcessor extends BaseYgdyProcessor {
                         douban_score0 = douban_score0.replaceAll("</font>", "");
                         String[] douban_score = douban_score0.trim().split("/");
                         if (douban_score.length > 0 && !TextUtils.isEmpty(douban_score[0])) {
-                            sourceDetail.setDoubanScore(Float.valueOf(StringUtil.disposeField(douban_score[0])));
+                            sourceDetail.setDoubanScore(StringUtil.disposeField(douban_score[0]));
                         }
                         if (douban_score.length > 1 && !TextUtils.isEmpty(douban_score[1])) {
                             sourceDetail.setDoubanIntro(StringUtil.disposeField(douban_score[1]));
@@ -360,7 +360,7 @@ public class YgdySourceDetailProcessor extends BaseYgdyProcessor {
                             douban_score1 = douban_score1.replaceAll("</font>", "");
                             String[] douban_score = douban_score1.trim().split("/");
                             if (douban_score.length > 0 && !TextUtils.isEmpty(douban_score[0])) {
-                                sourceDetail.setDoubanScore(Float.valueOf(StringUtil.disposeField(douban_score[0])));
+                                sourceDetail.setDoubanScore(StringUtil.disposeField(douban_score[0]));
                             }
                             if (douban_score.length > 1 && !TextUtils.isEmpty(douban_score[1])) {
                                 sourceDetail.setDoubanIntro(StringUtil.disposeField(douban_score[1]));
